refactor: extract symbol index into dedicated microcrate
