+++ crates/perl-lsp-tooling/src/performance.rs
@@ -7,9 +7,10 @@
 
 use moka::sync::Cache;
 use perl_parser_core::Node;
-use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
+
+pub use perl_symbol_index::SymbolIndex;
 
 /// Cache for parsed ASTs with TTL.
 ///
@@ -214,149 +215,6 @@ pub mod parallel {
     }
 
     use super::*;
-}
-
-/// Symbol index for fast lookups.
-///
-/// Supports both prefix and fuzzy matching using a trie and inverted index.
-pub struct SymbolIndex {
-    /// Trie structure for prefix matching
-    trie: SymbolTrie,
-    /// Inverted index for fuzzy matching
-    inverted_index: HashMap<String, Vec<String>>,
-}
-
-/// Trie data structure for efficient prefix matching
-struct SymbolTrie {
-    /// Child nodes indexed by character
-    children: HashMap<char, Box<SymbolTrie>>,
-    /// Symbols stored at this node
-    symbols: Vec<String>,
-}
-
-impl Default for SymbolIndex {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl SymbolIndex {
-    /// Create a new empty symbol index
-    pub fn new() -> Self {
-        Self { trie: SymbolTrie::new(), inverted_index: HashMap::new() }
-    }
-
-    /// Add a symbol to the index.
-    ///
-    /// Indexes the symbol for both prefix and fuzzy matching.
-    pub fn add_symbol(&mut self, symbol: String) {
-        // Add to trie for prefix matching
-        self.trie.insert(&symbol);
-
-        // Add to inverted index for fuzzy matching
-        let tokens = Self::tokenize(&symbol);
-        for token in tokens {
-            self.inverted_index.entry(token).or_default().push(symbol.clone());
-        }
-    }
-
-    /// Search symbols with prefix.
-    ///
-    /// Returns all symbols starting with the given prefix.
-    pub fn search_prefix(&self, prefix: &str) -> Vec<String> {
-        self.trie.search_prefix(prefix)
-    }
-
-    /// Fuzzy search symbols.
-    ///
-    /// Returns symbols matching any of the tokenized query words, sorted by relevance.
-    pub fn search_fuzzy(&self, query: &str) -> Vec<String> {
-        let tokens = Self::tokenize(query);
-        let mut results = HashMap::new();
-
-        for token in tokens {
-            if let Some(symbols) = self.inverted_index.get(&token) {
-                for symbol in symbols {
-                    *results.entry(symbol.clone()).or_insert(0) += 1;
-                }
-            }
-        }
-
-        // Sort by relevance (number of matching tokens)
-        let mut sorted: Vec<_> = results.into_iter().collect();
-        sorted.sort_by(|(_, a), (_, b)| b.cmp(a));
-
-        sorted.into_iter().map(|(symbol, _)| symbol).collect()
-    }
-
-    fn tokenize(s: &str) -> Vec<String> {
-        // Split on word boundaries and case changes
-        let mut tokens = Vec::new();
-        let mut current = String::new();
-        let mut prev_upper = false;
-
-        for ch in s.chars() {
-            if ch.is_uppercase() && !prev_upper && !current.is_empty() {
-                tokens.push(current.to_lowercase());
-                current = String::new();
-            }
-
-            if ch.is_alphanumeric() {
-                current.push(ch);
-                prev_upper = ch.is_uppercase();
-            } else if !current.is_empty() {
-                tokens.push(current.to_lowercase());
-                current = String::new();
-                prev_upper = false;
-            }
-        }
-
-        if !current.is_empty() {
-            tokens.push(current.to_lowercase());
-        }
-
-        tokens
-    }
-}
-
-impl SymbolTrie {
-    fn new() -> Self {
-        Self { children: HashMap::new(), symbols: Vec::new() }
-    }
-
-    fn insert(&mut self, symbol: &str) {
-        let mut node = self;
-
-        for ch in symbol.chars() {
-            node = node.children.entry(ch).or_insert_with(|| Box::new(SymbolTrie::new()));
-        }
-
-        node.symbols.push(symbol.to_string());
-    }
-
-    fn search_prefix(&self, prefix: &str) -> Vec<String> {
-        let mut node = self;
-
-        for ch in prefix.chars() {
-            match node.children.get(&ch) {
-                Some(child) => node = child,
-                None => return Vec::new(),
-            }
-        }
-
-        // Collect all symbols from this node and descendants
-        let mut results = Vec::new();
-        Self::collect_all(node, &mut results);
-        results
-    }
-
-    fn collect_all(node: &SymbolTrie, results: &mut Vec<String>) {
-        results.extend(node.symbols.clone());
-
-        for child in node.children.values() {
-            Self::collect_all(child, results);
-        }
-    }
 }
 
 #[cfg(test)]
