+++ crates/perl-symbol-index/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "perl-symbol-index"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Trie + inverted-index symbol search for Perl tooling"
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+repository.workspace = true
+homepage.workspace = true
+documentation = "https://docs.rs/perl-symbol-index"
+keywords = ["perl", "symbol", "index", "search", "lsp"]
+categories = ["text-editors", "development-tools"]
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
