+++ crates/perl-symbol-index/README.md
@@ -1,0 +1,7 @@
+# perl-symbol-index
+
+Single-responsibility symbol search index used by Perl LSP/tooling.
+
+Provides:
+- Prefix lookup using a trie
+- Tokenized fuzzy lookup using an inverted index
